Remove arbitrary restrictions on OP_RETURN by default
Any number of data-carrying OP_RETURN outputs are allowed, and the maximum data
carrier size is unlimited by default.

Rational: there's lots of ways for people to publish data in the Bitcoin chain,
and lots of data has been published.  There's no reason for us to place
artificial limits on this particular method.

-datacarrier and -datacarriersize are now totally ignored.